% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_Simple.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the csv file to be read. File should be in the same
working directory as the R code or an error will be generated.}
}
\value{
function returns a dataframe or an error if the file does not exist.
}
\description{
This function reads a csv file into R and returns a dataframe.
It is internal to this package.If the file name does not exist,
an error is generated and the function stopped.
}
\examples{
\dontrun{
dat <- fars_read("accident_2013.csv.bz2")
}

}
