% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_Simple.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of years(integers) such as \code{list(2013,2014,2015)}}
}
\value{
the function returns a dataframe with years as a variable and a 
summary of the accidents grouped by month.
}
\description{
This function takes as an argument a list composed of years(integers)
and returns a summary of accidents per month for a given year is a dataframe.
It calls the function \code{fars_read_years(years)}. If the year cannot be
found, then an error is generated.
}
\examples{
\dontrum{
fars_summarize_years(list(2013,2014,2015))
}

}
