% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_Simple.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year is an integer argument}
}
\value{
a string of the form "accident_year.csv.bz2"
}
\description{
This function generates a file namein the form "accident_%d.csv.bz2"
using the C wrapper sprintf where d = year (an integer). This
function is internal to this package. If year is not an integer,
then an error is generated.
}
\examples{
\dontrun{
file <- make_filename(2013)
}
}
